fixed hash issues and linked node shared ptrs. now we can assume that costumer and records are deleted automaticly.

+++ Customer.cpp
@@ -26,13 +26,13 @@ StatusType Customer::set_club_member(){
     return StatusType::SUCCESS;
 }
 
-int Customer::get_expense() const
+double Customer::get_expense() const
 {
     return expense;
 }
 
 /// \param amount adds the amount to the current value
-void Customer::update_expense(int amount){
+void Customer::update_expense(double amount){
     expense+=amount;
 }
